+++ Exposure.h
@@ -24,11 +24,11 @@ private:
 	uint8_t calcSSIndex(float exposureTime);
 
 public:
-	Exposure(int pin, float fstop = 8, uint16_t iso = 100, uint8_t ec = 0);
+	Exposure(int pin, float fstop = 8, uint16_t iso = 100, float ec = 0);
 	virtual ~Exposure();
 	uint16_t ISO;
 	float fStop;
-	uint8_t EC;
+	float EC;
 	float Expo;
 	float EVx;
 	float EV100;
